% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagi_bwm.R
\name{.dodaj_ograniczenie}
\alias{.dodaj_ograniczenie}
\title{Pomocnik budowania ograniczeń
Wykorzystane do budowania macierzy ograniczen}
\usage{
.dodaj_ograniczenie(ograniczenia, nowe_ograniczenie)
}
\description{
Pomocnik budowania ograniczeń
Wykorzystane do budowania macierzy ograniczen
}
\keyword{internal}
