% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{.motyw_wykresu}
\alias{.motyw_wykresu}
\title{Plot Fuzzy VIKOR Results}
\usage{
.motyw_wykresu()
}
\arguments{
\item{vikor_res}{Result object from fuzzy_vikor}
}
\description{
Visualizes the VIKOR results using a bubble chart.
}
\keyword{internal}
