% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.parsuj_skladnie_mcda}
\alias{.parsuj_skladnie_mcda}
\title{Wewnętrzny parser składni}
\usage{
.parsuj_skladnie_mcda(skladnia)
}
\description{
Zamienia tekst typu: "Kryterium =~ zmienna1 + zmienna2" na listę w R.
}
\keyword{internal}
