% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagi_bwm.R
\name{.sprawdz_spojnosc}
\alias{.sprawdz_spojnosc}
\title{Wewnętrzne sprawdzanie spójności
Z .waliduj_dane_bwm best_to_others = najlepsze_do_innych,
others_to_worst = inne_do_najgorszego, criteria_names = nazwy_kryteriow}
\usage{
.sprawdz_spojnosc(model)
}
\description{
Wewnętrzne sprawdzanie spójności
Z .waliduj_dane_bwm best_to_others = najlepsze_do_innych,
others_to_worst = inne_do_najgorszego, criteria_names = nazwy_kryteriow
}
\keyword{internal}
