% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_mcda.R
\name{.teoria_dominacji_multimoora}
\alias{.teoria_dominacji_multimoora}
\title{Wewnętrza funkcja MULTIMOORA - teoria dominacji}
\usage{
.teoria_dominacji_multimoora(r1, r2, r3)
}
\description{
Agreguje trzy wewnętrzne rankingi metody MULTIMOORA  (RS, RP, MF).
}
\keyword{internal}
