% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagi_bwm.R
\name{.waliduj_dane_bwm}
\alias{.waliduj_dane_bwm}
\title{Wewnętrzna walidacja danych}
\usage{
.waliduj_dane_bwm(najlepsze_do_innych, inne_do_najgorszego, nazwy_kryteriow)
}
\description{
Sprawdza, czy wektory porównań mają sens (długość, zakres 1-9). Korzysta z .wymus_bwm.
}
\keyword{internal}
