% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ranking.R
\name{fuzzy_meta_ranking}
\alias{fuzzy_meta_ranking}
\title{Rozmyty Meta-Ranking}
\usage{
fuzzy_meta_ranking(
  macierz_decyzyjna,
  typy_kryteriow,
  wagi = NULL,
  bwm_najlepsze = NULL,
  bwm_najgorsze = NULL,
  v = 0.5
)
}
\arguments{
\item{macierz_decyzyjna}{Rozmyta macierz decyzyjna.}

\item{typy_kryteriow}{Wektor typów kryteriów ("min", "max").}

\item{wagi}{Wektor wag (opcjonalny).}

\item{bwm_najlepsze, bwm_najgorsze}{Parametry BWM.}

\item{v}{Parametr VIKOR.}
}
\description{
Agreguje wyniki: VIKOR, MULTIMOORA.
}
