% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_mcda.R
\name{fuzzy_multimoora}
\alias{fuzzy_multimoora}
\title{Rozmyta Metoda MULTIMOORA}
\usage{
fuzzy_multimoora(
  macierz_decyzyjna,
  typy_kryteriow,
  wagi = NULL,
  bwm_kryteria,
  bwm_najlepsze,
  bwm_najgorsze
)
}
\arguments{
\item{macierz_decyzyjna}{Macierz rozmyta (m x 3n) w układzie (l1,m1,u1, l2,m2,u2, ...).}

\item{typy_kryteriow}{Wektor znakowy "max" (benefit) lub "min" (cost).}

\item{wagi}{Wektor wag (length = n_kryteriow lub length = 3*n_kryteriow).}

\item{bwm_kryteria}{Wektor z nazwami kryteriów (jeśli BWM).}

\item{bwm_najlepsze}{Wektor (best-to-others) dla BWM.}

\item{bwm_najgorsze}{Wektor (others-to-worst) dla BWM.}
}
\value{
`fuzzy_multimoora_res`
}
\description{
Implementacja metody Fuzzy MULTIMOORA. Składa się z:
1. Ratio System (RS)
2. Reference Point (RP)
3. Full Multiplicative Form (FMF)
Finalny ranking powstaje przez agregację Teorią Dominacji.
}
