% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_mcda.R
\name{fuzzy_vikor}
\alias{fuzzy_vikor}
\title{Implementacja Fuzzy VIKOR dla macierzy rozmytej (TFN)
Oblicza wskaźniki S (użyteczność grupy), R (indywidualny żal) oraz Q (indeks kompromisu).}
\usage{
fuzzy_vikor(
  macierz_decyzyjna,
  typy_kryteriow,
  v = 0.5,
  wagi = NULL,
  bwm_kryteria,
  bwm_najlepsze,
  bwm_najgorsze
)
}
\arguments{
\item{macierz_decyzyjna}{Macierz rozmyta (m x 3n) w układzie (l1,m1,u1, l2,m2,u2, ...).}

\item{typy_kryteriow}{Wektor znakowy "max" (benefit) lub "min" (cost).}

\item{v}{Waga strategii grupowej (domyślnie 0.5).}

\item{wagi}{Wektor wag (length = n_kryteriow lub length = 3*n_kryteriow).}

\item{bwm_kryteria}{Wektor z nazwami kryteriów (jeśli BWM).}

\item{bwm_najlepsze}{Wektor (best-to-others) dla BWM.}

\item{bwm_najgorsze}{Wektor (others-to-worst) dla BWM.}
}
\value{
`fuzzy_vikor_res`
}
\description{
Implements Fuzzy VIKOR with BWM integration. Returns an object for plotting.
}
