% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.R
\name{mcda_dane_surowe}
\alias{mcda_dane_surowe}
\title{Przykładowe dane do macierzy decyzyjnej MCDA}
\format{
Ramka danych (data frame) z 50 wierszami i 12 zmiennymi:
\describe{
  \item{EkspertID}{Identyfikator eksperta (1–5). Każdy ekspert ocenia wszystkie kraje.}
  \item{Alternatywa}{Nazwa kraju będącego alternatywą decyzyjną.}

  \item{war_oferty}{Liczba ofert pracy IT (zmienna ciągła).}
  \item{war_siec}{Jakość łącza internetowego (skala Likerta 1-5).}
  \item{war_elast}{Elastyczność godzin pracy (skala Likerta 1-5).}
  \item{koszt_mieszkanie}{Średni koszt zakwaterowania (EUR/miesiąc, zmienna ciągła).}
  \item{koszt_jedzenie}{Średni koszt jedzenia (EUR/miesiąc, zmienna ciągła).}
  \item{koszt_transport}{Średni koszt komunikacji lokalnej (EUR/miesiąc, zmienna ciągła).}
  \item{atr_kultura}{Atrakcyjność kultury (skala Likerta 1–7).}
  \item{atr_jezyk}{Ocena chęci i potrzeby nauki języka kraju (skala Likerta 1–7) Zawiera wartość 99 jako kod błędu/braku odpowiedzi.}
  \item{atr_obcy}{Łatwość zatrudnienia dla objokrajowców (skala Likerta 1–7).}
  \item{atr_zdrowie}{Dostępność usług zdrowotnych (skala Likerta 1–7).}
  \item{rozwoj_zaw}{Możliwość rozwoju zawodowego (skala Likerta 1–7).}
  \item{rozwoj_osob}{Możliwość rozwoju osobistego poza pracą (skala Likerta 1–7).}
  \item{roznica_czasu}{Różnica strefy czasowej względem Polski (0–2, zmienna dyskretna).}
}
}
\usage{
data(mcda_dane_surowe)
}
\description{
Zbiór danych zawierający symulowane odpowiedzi lub pomiary dla 10 alternatyw
ocenianych w ramach 10 zmiennych reprezentujących kryteria.
Zbiór przeznaczony do użycia z funkcją `przygotuj_dane_mcda()`.
}
