% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ranking.R
\name{oblicz_dominacje_ranking}
\alias{oblicz_dominacje_ranking}
\title{Teoria Dominacji dla Rankingu}
\usage{
oblicz_dominacje_ranking(rank_mat)
}
\description{
Funkcja pomocnicza do konsensusu.
}
\keyword{internal}
