% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ranking.R
\name{oblicz_ranking_dominacji}
\alias{oblicz_ranking_dominacji}
\title{Teoria Dominacji dla Rankingu}
\usage{
oblicz_ranking_dominacji(rank_mat)
}
\description{
Funkcja pomocnicza do wyznaczenia konsensusu na podstawie zasady większości.
}
\keyword{internal}
