% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fuzzy_vikor}
\alias{plot.fuzzy_vikor}
\title{Plot Fuzzy VIKOR Results}
\usage{
\method{plot}{fuzzy_vikor}(data)
}
\arguments{
\item{data}{An object of class `fuzzy_vikor_res` returned by the `fuzzy_vikor()` function.}
}
\value{
A ggplot object.
}
\description{
Visualizes the VIKOR results using a bubble chart.
}
