% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja.R
\name{tabela_apa}
\alias{tabela_apa}
\title{Generowanie Tabeli APA}
\usage{
tabela_apa(x, tytul = NULL)
}
\arguments{
\item{x}{Obiekt wynikowy z funkcji pakietu (np. `fuzzy_vikor_res`).}

\item{tytul}{Opcjonalny tytuł tabeli.}
}
\value{
Obiekt klasy `flextable` gotowy do druku lub zapisu do Worda.
}
\description{
Funkcja przekształca wyniki analizy MCDA w sformatowaną tabelę
zgodną ze standardem APA, gotową do publikacji w Wordzie.
}
\details{
------------- Tabele APA ----------------------
}
