% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.R
\name{mcda_dane_surowe}
\alias{mcda_dane_surowe}
\title{Przykładowe dane do macierzy decyzyjnej MCDA}
\format{
Ramka danych (data frame) z 50 wierszami i 12 zmiennymi:
\describe{

  \item{EkspertID}{Identyfikator eksperta (1–5). Każdy ekspert ocenia wszystkie kraje.}

  \item{Alternatywa}{Nazwa kraju będącego alternatywą decyzyjną.}

  \item{oferty_IT}{Liczba ofert pracy IT (zmienna ciągła).}

  \item{wynagrodzenie}{Poziom wynagrodzenia (EUR/miesiąc, zmienna ciągła).}

  \item{koszt_mieszkania}{Średni koszt wynajmu mieszkania (EUR/miesiąc, zmienna ciągła).}

  \item{odleglosc_km}{Odległość geograficzna od Polski (km, zmienna ciągła).}

  \item{dostep_jedzenie_uslugi}{Liczba punktów usług/żywienia w okolicy (zmienna ciągła).}

  \item{roznica_czasu}{Różnica strefy czasowej względem Polski (0–2, zmienna dyskretna).}

  \item{zatrudnienie_obcy}{Łatwość zatrudnienia dla objokrajowców (skala Likerta 1–7).}

  \item{komunikacja_lokalna}{Ocena jakości komunikacji lokalnej (skala Likerta 1–7).}

  \item{rozwoj}{Możliwość rozwoju zawodowego (skala Likerta 1–7). Zawiera wartość 99 jako kod błędu/braku odpowiedzi.}

  \item{przyjaznosc_kultury}{Ocena przyjazności kultury (skala Likerta 1–7).}

}
}
\usage{
data(mcda_dane_surowe)
}
\description{
Zbiór danych zawierający symulowane odpowiedzi lub pomiary dla 10 alternatyw
ocenianych w ramach 10 zmiennych reprezentujących kryteria.
Zbiór przeznaczony do użycia z funkcją `przygotuj_dane_mcda()`.
}
