% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja.R
\name{plot.fuzzy_vikor_res}
\alias{plot.fuzzy_vikor_res}
\title{Wizualizacja Fuzzy VIKOR}
\usage{
\method{plot}{fuzzy_vikor_res}(x, ...)
}
\arguments{
\item{x}{Wynik z fuzzy_vikor.}

\item{...}{Dodatkowe argumenty przekazywane do funkcji plot.}
}
\description{
Wizualizacja Fuzzy VIKOR
}
